+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Added
+- Added mobile app version of "My certificates"
+
 ## 5.0.1 - 2025-06-10
 ### Fixed
 - Fixed an issue where course navigation was not working correctly when viewing a certificate